Login is now fully functionnal Added the possibility to add/update/delete medias linked to tricks Proper display of tricks on homepage Proper display of medias on trick's page and update tricks Added choice for a media to be either a picture or a video Added the possibility to choose the main picture of a trick Added the possibility to add comments (not yet linked to a trick)

+++ src/Controller/STController.php
@@ -6,12 +6,17 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use App\Entity\Utilisateur;
 use App\Entity\Figure;
 use App\Entity\Media;
+use App\Entity\Message;
+use App\Form\Type\MessageType;
 use App\Form\Type\UtilisateurType;
+use App\Form\Type\UpdateFigureType;
 use App\Form\Type\FigureType;
+use App\Form\Type\MediaType;
 use App\Form\Type\LoginType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
+use Doctrine\Common\Collections\ArrayCollection;
 
 class STController extends AbstractController{
     
@@ -23,12 +28,25 @@ class STController extends AbstractController{
         return $this->render('index.html.twig',['figures'=>$figures]);
     }
     
-    public function afficherFigure($nomFigure){
+    public function afficherFigure($nomFigure,Request $request){
         
         $entityManager = $this->getDoctrine()->getManager();
         $figure= $entityManager->getRepository(Figure::class)->findOneBy(['nom' => $nomFigure]);;
-       
-        return $this->render('figure.html.twig',['figure'=>$figure]);
+        $medias=$figure->getMedias();
+        //$messages=$figure->getMessages();
+        
+        $message = new Message;
+        $form = $this->createForm(MessageType::class, $message);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $message = $form->getData();
+            $message->setUtilisateur($this->getUser());
+            $message->setDate(new \DateTime('now'));
+            $entityManager->persist($message);
+            $entityManager->flush();
+            return $this->redirectToRoute('figure',['nomFigure'=>$nomFigure]);
+        }
+        return $this->render('figure.html.twig',['figure'=>$figure, 'medias'=>$medias, 'form' => $form->createView()]);
     }
     
     //Affiche la page d'inscription
@@ -41,7 +59,7 @@ class STController extends AbstractController{
         if ($form->isSubmitted() && $form->isValid()) {
             $utilisateur = $form->getData();
             $notHashedPassword=$utilisateur->getMotDePasse();
-            $utilisateur->setMotDePasse(password_hash($notHashedPassword, PASSWORD_DEFAULT));
+            $utilisateur->setPassword(password_hash($notHashedPassword, PASSWORD_DEFAULT));
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($utilisateur);
             $entityManager->flush();
@@ -86,10 +104,10 @@ class STController extends AbstractController{
         $figure = new Figure();
         $form = $this->createForm(FigureType::class, $figure);
         $form->handleRequest($request);
+        
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $figure = $form->getData();
-            //EN ATTENDANT DE DEBUGGER LOGIN -> TOUT CREEE PAR USER 1
             $figure->setUtilisateur($user);
             foreach($figure->getMedias() as $media) {
                 if(!$figure->getMedias()->contains($media))
@@ -97,6 +115,8 @@ class STController extends AbstractController{
                     $figure->addMedia($media);
                 }
             }
+            $dateCreation= new \DateTime('now');
+            $figure->setDateCreation($dateCreation);
             $entityManager->persist($figure);
             $entityManager->flush();
             return $this->redirectToRoute('index');
@@ -106,12 +126,37 @@ class STController extends AbstractController{
         ]);
     }
     
-    public function modifierFigure($idFigure){
-    
+    public function modifierFigure($idFigure,Request $request){
+        
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $entityManager = $this->getDoctrine()->getManager();
+        $figure = $entityManager->getRepository(Figure::class)->find($idFigure);
+        $user = $this->getUser();
+        $medias = new arrayCollection();
+        foreach($figure->getMedias() as $media){
+            $medias->add($media);
+        }
+        
+        $form = $this->createForm(UpdateFigureType::class, $figure);
+        
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $figure->setDateDerniereModification(new \DateTime('now'));
+            $figure->setDernierUtilisateurModification($user);
+            $entityManager->persist($figure);
+            $entityManager->flush();
+            return $this->redirectToRoute('index');
+        }
+        return $this->render('modifierFigureForm.html.twig', [
+            'form' => $form->createView(),
+            'figure'=> $figure,
+            'medias'=>$medias
+        ]);
     }
     
     public function supprimerFigure($idFigure){
         
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $entityManager = $this->getDoctrine()->getManager();
         $figure = $entityManager->getRepository(Figure::class)->find($idFigure);
         $entityManager->remove($figure);
@@ -119,4 +164,71 @@ class STController extends AbstractController{
         
         return $this->redirectToRoute('index');
     }
+    
+    public function ajoutMedia($idFigure,Request $request){
+        
+        // usually you'll want to make sure the user is authenticated first
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+       $entityManager = $this->getDoctrine()->getManager();
+        $figure = $entityManager->getRepository(Figure::class)->find($idFigure);
+   
+        $media = new Media();
+        $form = $this->createForm(MediaType::class, $media);
+        $form->handleRequest($request);
+        
+        if ($form->isSubmitted() && $form->isValid()) {
+            $media = $form->getData();
+            $media->setFigure($figure);
+            $figure->addMedia($media);
+            
+            $entityManager->persist($media);
+            $entityManager->persist($figure);
+            $entityManager->flush();
+            
+            return $this->redirect('/SnowTricks/public/modifier-une-figure/'.$idFigure.'');
+        }
+        return $this->render('ajoutMediaForm.html.twig', [
+            'form' => $form->createView(),
+        ]);
+    }
+    
+    public function ModifierMedia($idMedia,$idFigure,Request $request){
+        
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $entityManager = $this->getDoctrine()->getManager();
+        $figure= $entityManager->getRepository(Figure::class)->find($idFigure);
+        $media = $entityManager->getRepository(Media::class)->find($idMedia);
+        $form = $this->createForm(MediaType::class, $media);
+        $form->handleRequest($request);
+        
+        if ($form->isSubmitted() && $form->isValid()) {
+            $media = $form->getData();
+            $figure->setDateDerniereModification(new \DateTime('now'));
+            
+            $entityManager->persist($media);
+            $entityManager->persist($figure);
+            $entityManager->flush();
+            
+            return $this->redirect('/SnowTricks/public/modifier-une-figure/'.$idFigure.'');
+        }
+        return $this->render('modifierMediaForm.html.twig', [
+            'form' => $form->createView(),
+            'media'=>$media,
+        ]);
+        
+        return $this->redirect('/SnowTricks/public/modifier-une-figure/'.$idFigure.'');
+    }
+    
+    public function supprimerMedia($idMedia,$idFigure){
+        
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $entityManager = $this->getDoctrine()->getManager();
+        $media = $entityManager->getRepository(Media::class)->find($idMedia);
+        $entityManager->remove($media);
+        $entityManager->flush();
+        
+        return $this->redirect('/SnowTricks/public/modifier-une-figure/'.$idFigure.'');
+    }
+    
 }
